generate kernels for codegend out= operators
[ghstack-poisoned]
