+++ aten/src/ATen/templates/CompositeViewCopyKernels.cpp
@@ -28,10 +28,23 @@ std::vector<at::Tensor> clone_arg(const at::TensorList& t_list) {
     return out;
 }
 
+void copy_arg(const at::Tensor& dst, const at::Tensor& src) {
+    dst.copy_(src);
+}
+
+void copy_arg(const at::TensorList& dst, const at::TensorList& src) {
+    TORCH_INTERNAL_ASSERT(dst.size() == src.size());
+    for (const auto& i : c10::irange(dst.size())) {
+        dst[i].copy_(src[i]);
+    }
+}
+
 
 ${CompositeViewCopyKernel_Definitions}
 
 ${GeneratedCompositeFunctional_Definitions}
+
+${GeneratedCompositeOut_Definitions}
 
 } // namespace native
 } // namespace at
